+++ lib/llamacpp_tools.dart
@@ -1,4 +1,4 @@
-export 'src/docker_builder.dart';
-export 'src/github_downloader.dart';
+export 'src/docker/docker_builder.dart';
+export 'src/llamacpp_github.dart';
 export 'src/llamacpp_dir.dart';
 export 'src/llamaserver_process.dart';